Fixed container not found in cache while updating
The bug was from container caching, during update, the loader method
`parse_sub_matrix` was called twice, which means the container cache
lookup `hierarchy.container_from_id_path` was called twice as well,
which means, the cached container was able to matched in first call,
but the next call, which was about to update the matrix, could not
found the container.

To resolve this, we gather all container ids that needs to be resolved,
and resolve them in one batch.